+++ Example/Tests/BLECommunicatorTests.swift
@@ -88,9 +88,8 @@ class BLECommunicatorTests: XCTestCase {
         }
         intervals.sort(by: { $0 > $1 })
 
-        let mockPeripheral = BLEScanner().sidPeripheral
         for interval in intervals {
-            let mockSid = SID(sidID: mockSidId, peripheral: mockPeripheral, discoveryDate: reference.addingTimeInterval(-interval) as Date, isConnected: false, rssi: 0)
+            let mockSid = SID(sidID: mockSidId, peripheral: nil, discoveryDate: reference.addingTimeInterval(-interval) as Date, isConnected: false, rssi: 0)
             communicator.transferDidDiscoveredSidId(transporter, newSid: mockSid)
         }
 
@@ -119,11 +118,10 @@ class BLECommunicatorTests: XCTestCase {
      Mocked data used for testing purpose
      */
     func refillMockSids() {
-        let mockPeripheral = BLEScanner().sidPeripheral
-        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "bb28d13fdcab416b85b7cec28c26add7", peripheral: mockPeripheral, discoveryDate: Date().addingTimeInterval(-0.8) as Date, isConnected: false, rssi: 0))
-        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "550e8400e29b11d4a716446655440003", peripheral: mockPeripheral, discoveryDate: Date().addingTimeInterval(-1.8) as Date, isConnected: false, rssi: 0))
-        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "1a1092e99f824187af92d92029b28cdc", peripheral: mockPeripheral, discoveryDate: Date().addingTimeInterval(-2.5) as Date, isConnected: false, rssi: 0))
-        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "2c6088153bc7434f9c2b2e3272596adc", peripheral: mockPeripheral, discoveryDate: Date().addingTimeInterval(-3.0) as Date, isConnected: false, rssi: 0))
-        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "250bf2429d8c4f2896e2030dfe601bd8", peripheral: mockPeripheral, discoveryDate: Date().addingTimeInterval(-4.8) as Date, isConnected: false, rssi: 0))
+        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "bb28d13fdcab416b85b7cec28c26add7", peripheral: nil, discoveryDate: Date().addingTimeInterval(-0.8) as Date, isConnected: false, rssi: 0))
+        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "550e8400e29b11d4a716446655440003", peripheral: nil, discoveryDate: Date().addingTimeInterval(-1.8) as Date, isConnected: false, rssi: 0))
+        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "1a1092e99f824187af92d92029b28cdc", peripheral: nil, discoveryDate: Date().addingTimeInterval(-2.5) as Date, isConnected: false, rssi: 0))
+        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "2c6088153bc7434f9c2b2e3272596adc", peripheral: nil, discoveryDate: Date().addingTimeInterval(-3.0) as Date, isConnected: false, rssi: 0))
+        communicator.transferDidDiscoveredSidId(transporter, newSid: SID(sidID: "250bf2429d8c4f2896e2030dfe601bd8", peripheral: nil, discoveryDate: Date().addingTimeInterval(-4.8) as Date, isConnected: false, rssi: 0))
     }
 }